refactor: remove dormant ai sdk runtime
